Closer to being functional
Most of the updated code is straight from one of the rlbot-rust examples. Once I get it working with that code I can implement the packet converter and tie that to my own setup. It will be smooth sailing form then on. The current error I am getting is below and I will figure it out in the morning.
Error: StringError("not launched by framework")