Further Implementation of Own Code
Now that the bot is functioning, I am implementing my own code. Still need to finish drawing data from the GameTickPacket to put into my own packet form and finish the output by linking the CurrentState enums to the different state structures but it is late so I will pick up on this tomorrow.